+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl.java
@@ -89,7 +89,7 @@ public class NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl im
 
   public NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl(ClientState clientState, HasMetadata item) {
     this(HasMetadataOperationsImpl.defaultContext(clientState).withItem(item), new NamespaceVisitOperationContext());
-    handlerOf(item, new BaseClient(clientState)); // validate the handler
+    handlerOf(item, this.context); // validate the handler
   }
 
   @Override
@@ -126,7 +126,7 @@ public class NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl im
 
   @Override
   public VisitFromServerGetWatchDeleteRecreateWaitApplicable<HasMetadata> inNamespace(String explicitNamespace) {
-    HasMetadata item = acceptVisitors(get(), Collections.emptyList(), explicitNamespace);
+    HasMetadata item = acceptVisitors(get(), Collections.emptyList(), explicitNamespace, this.context);
     return newInstance(context.withItem(item).withNamespace(explicitNamespace), namespaceVisitOperationContext.withExplicitNamespace(explicitNamespace));
   }
 
@@ -142,7 +142,7 @@ public class NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl im
 
   @Override
   public VisitFromServerGetWatchDeleteRecreateWaitApplicable<HasMetadata> accept(Visitor... visitors) {
-    HasMetadata item = acceptVisitors(get(), Arrays.asList(visitors), namespaceVisitOperationContext.getExplicitNamespace());
+    HasMetadata item = acceptVisitors(get(), Arrays.asList(visitors), namespaceVisitOperationContext.getExplicitNamespace(), this.context);
     return newInstance(context.withItem(item), namespaceVisitOperationContext);
   }
 
@@ -183,7 +183,7 @@ public class NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl im
   
   Resource<HasMetadata> getResource() {
     HasMetadata meta = (HasMetadata) context.getItem();
-    ResourceHandler<HasMetadata, ?> handler = handlerOf(meta, new BaseClient(this.context));
+    ResourceHandler<HasMetadata, ?> handler = handlerOf(meta, context);
     HasMetadataOperation<HasMetadata, ?, Resource<HasMetadata>> operation = handler.operation(context, null);
     return operation.newInstance(context).inNamespace(KubernetesResourceUtil.getNamespace(meta)).withName(KubernetesResourceUtil.getName(meta));
   }
@@ -240,8 +240,8 @@ public class NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl im
     return getResource().accept(function);
   }
   
-  static HasMetadata acceptVisitors(HasMetadata item, List<Visitor> visitors, String explicitNamespace) {
-    ResourceHandler<HasMetadata, ?> h = handlerOf(item, null);
+  static HasMetadata acceptVisitors(HasMetadata item, List<Visitor> visitors, String explicitNamespace, OperationContext context) {
+    ResourceHandler<HasMetadata, ?> h = handlerOf(item, context);
     VisitableBuilder<HasMetadata, ?> builder = h.edit(item);
 
     //Let's apply any visitor that might have been specified.
@@ -254,8 +254,8 @@ public class NamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableImpl im
     return builder.build();
   }
 
-  static <T extends HasMetadata, V extends VisitableBuilder<T, V>> ResourceHandler<T, V> handlerOf(T item, BaseClient client) {
-    ResourceHandler<T, V> result = Handlers.get(item, client);
+  static <T extends HasMetadata, V extends VisitableBuilder<T, V>> ResourceHandler<T, V> handlerOf(T item, OperationContext context) {
+    ResourceHandler<T, V> result = Handlers.get(item, new BaseClient(context));
     if (result == null) {
       throw new KubernetesClientException("Could not find a registered handler for item: [" + item + "].");
     }